info_copping_app/requirements.txt
flask~=2.0.2
flask_sqlalchemy
flask_migrate
flask_login
flask_mail
flask_moment
flask_wtf
werkzeug~=2.0.2
WTForms~=3.0.0
sqlalchemy~=1.4.27
psycopg2-binary
alembic~=1.7.5
email_validator
python-dotenv